Added an extension interface object
Currently the extension interface object is used for pub/sub.  This is needed in the flow-server extension to add hooks for
the socket connecting and disconnecting.

The extension loader will fall back to not including the extension
interface so that extensions that haven't updated their create
function will continue to work.