Pull configuration values from more places
To better support running the server from a generic Docker container, allow the
location of the settings file to be specified using an environment variable, and
also allow individual settings to be passed in using environment variables.

Add a bit of documentation to the sample config file, and modify the prep script
to generate random strings for the config values that need them.

Existing `settings/config.py` files should continue to work as before, though it
is no longer necessary to create `settings/__init__.py` because the file is now
imported as a standalone file rather than a module.