PEP8 E221: Single space before operator
This PEP8 rule was only violated by the model definitions. Since removing the
alignment of the equal signs messed up the alignment of the comments, I also
moved most of the comments to the actual column definitions; SQLAlchemy will
include them as column comments on PostgreSQL when it creates tables on a fresh
database.

There were a few PEP8 rules listed as exceptions in the README file that the
code wasn't actually violating, so I removed those too.