Fix Flake8 issues; add to CI suite
Now that pycodestyle is clean, up the ante by using Flake8, which includes
pycodestyle in addition to other checks.

The code changes are mostly unused imports and local variables, but Flake8 found
an actual bug: the wrong user ID was being used in `KeyList.post()`. It also found
a call to `xrange()` which was removed in Python 3.

In a couple of places, the unused variables were used by commented-out debugging
code; I removed the commented-out code since it's trivial to add locally if
needed.

To fix the unused imports in `app_run.py`, I added a dummy function that
references a couple of the symbols. Unclear to me why, but referencing just two of
the imports in code made multiple code checkers (Flake8, pylint, PyCharm) treat
the remaining imports as used.